Clean up tests.
