+++ internal/grid/grid_test.go
@@ -22,9 +22,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"runtime"
 	"strconv"
@@ -35,56 +32,18 @@ import (
 	"github.com/minio/minio/internal/logger/target/testlogger"
 )
 
-func startServer(t testing.TB, listener net.Listener, handler http.Handler) (server *httptest.Server) {
-	t.Helper()
-	server = httptest.NewUnstartedServer(handler)
-	server.Config.Addr = listener.Addr().String()
-	server.Listener = listener
-	server.Start()
-	t.Cleanup(server.Close)
-	t.Log("Started server on", server.Config.Addr, "URL:", server.URL)
-	return server
-}
-
-func shutdownManagers(t testing.TB, servers ...*Manager) {
-	t.Helper()
-	for _, s := range servers {
-		s.debugMsg(debugShutdown)
-	}
-	for _, s := range servers {
-		s.debugMsg(debugWaitForExit)
-	}
-}
-
 func TestSingleRoundtrip(t *testing.T) {
 	defer testlogger.T.SetLogTB(t)()
-	hosts, listeners, _ := getHosts(2)
-	dialer := &net.Dialer{
-		Timeout: 5 * time.Second,
-	}
 	errFatal := func(err error) {
 		t.Helper()
 		if err != nil {
 			t.Fatal(err)
 		}
 	}
-	wrapServer := func(handler http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			t.Logf("Got a %s request for: %v", r.Method, r.URL)
-			handler.ServeHTTP(w, r)
-		})
-	}
-	// We fake a local and remote server.
-	localHost := hosts[0]
-	remoteHost := hosts[1]
-	local, err := NewManager(context.Background(), ManagerOptions{
-		Dialer:      dialer.DialContext,
-		Local:       localHost,
-		Hosts:       hosts,
-		AuthRequest: dummyRequestValidate,
-		AddAuth:     func(aud string) string { return aud },
-	})
+	grid, err := SetupTestGrid(2)
 	errFatal(err)
+	remoteHost := grid.Hosts[1]
+	local := grid.Managers[0]
 
 	// 1: Echo
 	errFatal(local.RegisterSingleHandler(handlerTest, func(payload []byte) ([]byte, *RemoteErr) {
@@ -98,20 +57,7 @@ func TestSingleRoundtrip(t *testing.T) {
 		return nil, &err
 	}))
 
-	remote, err := NewManager(context.Background(), ManagerOptions{
-		Dialer:      dialer.DialContext,
-		Local:       remoteHost,
-		Hosts:       hosts,
-		AuthRequest: dummyRequestValidate,
-		AddAuth:     func(aud string) string { return aud },
-	})
-	errFatal(err)
-	defer shutdownManagers(t, local, remote)
-
-	localServer := startServer(t, listeners[0], wrapServer(local.Handler()))
-	defer localServer.Close()
-	remoteServer := startServer(t, listeners[1], wrapServer(remote.Handler()))
-	defer remoteServer.Close()
+	remote := grid.Managers[1]
 
 	// 1: Echo
 	errFatal(remote.RegisterSingleHandler(handlerTest, func(payload []byte) ([]byte, *RemoteErr) {
@@ -184,34 +130,18 @@ func TestSingleRoundtrip(t *testing.T) {
 }
 
 func TestSingleRoundtripGenerics(t *testing.T) {
-	defer testlogger.T.SetErrorTB(t)()
-	hosts, listeners, _ := getHosts(2)
-	dialer := &net.Dialer{
-		Timeout: 5 * time.Second,
-	}
+	defer testlogger.T.SetLogTB(t)()
 	errFatal := func(err error) {
 		t.Helper()
 		if err != nil {
 			t.Fatal(err)
 		}
 	}
-	wrapServer := func(handler http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			t.Logf("Got a %s request for: %v", r.Method, r.URL)
-			handler.ServeHTTP(w, r)
-		})
-	}
-	// We fake a local and remote server.
-	localHost := hosts[0]
-	remoteHost := hosts[1]
-	local, err := NewManager(context.Background(), ManagerOptions{
-		Dialer:      dialer.DialContext,
-		Local:       localHost,
-		Hosts:       hosts,
-		AddAuth:     func(aud string) string { return aud },
-		AuthRequest: dummyRequestValidate,
-	})
+	grid, err := SetupTestGrid(2)
 	errFatal(err)
+	remoteHost := grid.Hosts[1]
+	local := grid.Managers[0]
+	remote := grid.Managers[1]
 
 	// 1: Echo
 	h1 := NewSingleHandler[*testRequest, *testResponse](handlerTest, func() *testRequest {
@@ -242,29 +172,12 @@ func TestSingleRoundtripGenerics(t *testing.T) {
 	errFatal(h1.Register(local, handler1))
 	errFatal(h2.Register(local, handler2))
 
-	remote, err := NewManager(context.Background(), ManagerOptions{
-		Dialer:      dialer.DialContext,
-		Local:       remoteHost,
-		Hosts:       hosts,
-		AddAuth:     func(aud string) string { return aud },
-		AuthRequest: dummyRequestValidate,
-	})
-
-	errFatal(err)
-	defer shutdownManagers(t, local, remote)
-
 	errFatal(h1.Register(remote, handler1))
 	errFatal(h2.Register(remote, handler2))
-
-	localServer := startServer(t, listeners[0], wrapServer(local.Handler()))
-	defer localServer.Close()
-	remoteServer := startServer(t, listeners[1], wrapServer(remote.Handler()))
-	defer remoteServer.Close()
 
 	// local to remote connection
 	remoteConn := local.Connection(remoteHost)
 	const testPayload = "Hello Grid World!"
-	remoteConn.WaitForConnect(context.Background())
 
 	start := time.Now()
 	req := testRequest{Num: 1, String: testPayload}
@@ -289,60 +202,23 @@ func TestSingleRoundtripGenerics(t *testing.T) {
 
 func TestStreamSuite(t *testing.T) {
 	defer testlogger.T.SetErrorTB(t)()
-	hosts, listeners, _ := getHosts(2)
-	dialer := &net.Dialer{
-		Timeout: 1 * time.Second,
-	}
 	errFatal := func(err error) {
 		t.Helper()
 		if err != nil {
 			t.Fatal(err)
 		}
 	}
-	wrapServer := func(handler http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			t.Logf("Got a %s request for: %v", r.Method, r.URL)
-			handler.ServeHTTP(w, r)
-		})
-	}
-	// We fake a local and remote server.
-	localHost := hosts[0]
-	remoteHost := hosts[1]
-
-	local, err := NewManager(context.Background(), ManagerOptions{
-		Dialer:      dialer.DialContext,
-		Local:       localHost,
-		Hosts:       hosts,
-		AddAuth:     func(aud string) string { return aud },
-		AuthRequest: dummyRequestValidate,
-	})
-
+	grid, err := SetupTestGrid(2)
 	errFatal(err)
+	t.Cleanup(grid.Cleanup)
 
-	remote, err := NewManager(context.Background(), ManagerOptions{
-		Dialer:      dialer.DialContext,
-		Local:       remoteHost,
-		Hosts:       hosts,
-		AddAuth:     func(aud string) string { return aud },
-		AuthRequest: dummyRequestValidate,
-	})
-	errFatal(err)
-	defer shutdownManagers(t, local, remote)
-
-	localServer := startServer(t, listeners[0], wrapServer(local.Handler()))
-	remoteServer := startServer(t, listeners[1], wrapServer(remote.Handler()))
+	local := grid.Managers[0]
+	localHost := grid.Hosts[0]
+	remote := grid.Managers[1]
+	remoteHost := grid.Hosts[1]
 
 	connLocalToRemote := local.Connection(remoteHost)
-	err = connLocalToRemote.WaitForConnect(context.Background())
-	errFatal(err)
-
 	connRemoteLocal := remote.Connection(localHost)
-	err = connRemoteLocal.WaitForConnect(context.Background())
-	errFatal(err)
-	defer func() {
-		localServer.Close()
-		remoteServer.Close()
-	}()
 
 	t.Run("testStreamRoundtrip", func(t *testing.T) {
 		defer timeout(5 * time.Second)()
